backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langchain==0.1.0
langgraph==0.0.10
langchain-openai==0.0.5
redis==5.0.1
pydantic==2.6.0
httpx==0.26.0
python-dotenv==1.0.0
regex
google-generativeai
python-multipart
